code pull

+++ app/src/main/java/com/apollopharmacy/vishwam/ui/home/apna/model/SurveyDetailsList.kt
@@ -302,11 +302,11 @@ class SurveyDetailsList : Serializable {
 
         @SerializedName("building_age")
         @Expose
-        var buildingAge: Double? = null
+        var buildingAge: Int? = null
 
         @SerializedName("bldg_age_in_month")
         @Expose
-        var bldgAgeInMonth: Double? = null
+        var bldgAgeInMonth: Int? = null
 
         @SerializedName("morning_from")
         @Expose
@@ -594,7 +594,7 @@ class SurveyDetailsList : Serializable {
             return this
         }
 
-        fun withBuildingAge(buildingAge: Double?): Data {
+        fun withBuildingAge(buildingAge: Int?): Data {
             this.buildingAge = buildingAge
             return this
         }
